+++ scripts/settings.js
@@ -6,7 +6,6 @@ game.settings.register("campaign-codex", "itemPricePath", {
   hint: "Path to item price value (e.g., 'system.price.value' for D&D5e). Leave blank to use detected system.",
   scope: "world",
   config: true,
-    requiresReload: true,
   type: String,
   default: ""
 });
@@ -15,7 +14,6 @@ game.settings.register("campaign-codex", "itemDenominationPath", {
   hint: "Path to item currency denomination (e.g., 'system.price.denomination' for D&D5e). Leave blank to use detected system.",
   scope: "world",
   config: true,
-    requiresReload: true,
   type: String,
   default: ""
 });
@@ -24,7 +22,6 @@ game.settings.register("campaign-codex", "itemDenominationOverride", {
   hint: "Currency Denomination (e.g., credits, gp, units). Leave blank to use Item Currency Path or detected system. Campaign Codex will automatically detect the value and denomination for items based on the active game system (e.g., Dungeons and Dragons 5e, Pathfinder 2e, Starfinder, Savage Worlds Adventure Edition, Pathfinder 1e, Old-School Essentials and Daggerheart).",
   scope: "world",
   config: true,
-    requiresReload: true,
   type: String,
   default: ""
 });
@@ -33,16 +30,23 @@ game.settings.register("campaign-codex", "hideBaseCost", {
     hint: "If you don't want the base cost visible, or the system doesn't have a price path you can set this to hide the column in the inventory tab. Close and reopen sheets to update.",
     scope: "world",
     config: true,
-    // requiresReload: true,
     type: Boolean,
     default: false
+  });
+game.settings.register("campaign-codex", "sortCardsAlpha", { 
+    name: "Automatically sort cards alphabetically",                  
+    hint: "If On, cards will be displayed A-Z",
+    scope: "world",                                 
+    config: true,   
+    type: Boolean,
+    default: true,                                 
   });
 game.settings.register("campaign-codex", "runonlyonce", { 
     name: "Welcome Message - Disabled",                  
     hint: "If On, you won't see the Welcome message",
     scope: "world",                                 
     config: true,   
-      requiresReload: true,                                
+    requiresReload: true,                                
     type: Boolean,
     default: false,                                 
   });
@@ -51,7 +55,7 @@ game.settings.register("campaign-codex", "runonlyonce", {
     hint: "Automatically create and organise Campaign Codex journals in folders",
     scope: "world",
     config: true,
-      requiresReload: true,
+    requiresReload: true,
     type: Boolean,
     default: true
   });